Random Forest Models Added
This folder contains random forest models for frost classification. This file contains the baseline random forest used for comparisons between other models. The file also contains the Random Forest + Spatial Function, as well as tuning of the lambda parameter in the spatial function. Additionally, this code includes the get_cost function used to act as the cost-value matrix for model cost-value as discussed in novelty from paper.