Added data files for user convenience
In cases where users cannot run the data extraction or do not want to due to long computation times, they may use the files within this folder for easy usage of model code. Please note that GitHub does not support files larger than 25 MB. The combined dataset needed for model construction is larger than this size. In order to run models, a new data file must be made that concatenates all of these files together. This must be done by the user, or the data extraction code must be used.